refactor(#1459): split io/save_data + io/load_data helpers into private siblings (Path D)
Phase 2 of the backend god-file refactor umbrella (closed #1427).

Extract the module-level helper functions in
``src/scistudio/blocks/io/savers/save_data.py`` (1247 LOC) and
``src/scistudio/blocks/io/loaders/load_data.py`` (1217 LOC) into private
sibling modules under each sub-package. The ``SaveData`` / ``LoadData``
classes and their six ``_save_*`` / ``_load_*`` dispatch functions stay
in the original files; only the lower-level helpers move.

Mirror-symmetric layout (issue #1459 target):

  savers/                                loaders/
    __init__.py    (unchanged)             __init__.py    (unchanged)
    save_data.py   (618 LOC)               load_data.py   (733 LOC)
    _capability.py (FormatCapability+      _capability.py (FormatCapability+
                    _SAVE_EXTENSION_MAP)                    _LOAD_EXTENSION_MAP)
    _helpers.py    (path/pickle/coerce)    _helpers.py    (path/pickle/MIME/text)
    _streaming.py  (ADR-031 Phase 3
                    zarr-copy + CSV/PQ
                    row-group writers)

§C9 compliance (ADR-028 Addendum 1, "private functions, not helper
classes") is satisfied verbatim. Grep ``^class `` over the seven changed
files prints only ``SaveData`` and ``LoadData`` — every extracted
sibling holds underscore-prefixed module-level functions and module
constants only, no classes.

Public import surface preserved. ``from scistudio.blocks.io.savers
import SaveData`` and ``from scistudio.blocks.io.loaders import
LoadData`` continue to resolve unchanged. The legacy private symbols
(``_SAVE_CAPABILITIES``, ``_LOAD_EXTENSION_MAP``, ``_supported_save_extensions``,
``logger``, ``_LOGGER``, etc.) that tests historically imported directly
from ``save_data`` / ``load_data`` are re-exported via explicit
sibling-module imports at the top of each class file (with ``noqa: F401``
where ruff would otherwise object). A new structural test in
``tests/blocks/io/test_io_data_helpers_layout.py`` (46 cases) guards
both the §C9 invariant (AST-walks each sibling and asserts zero
``class`` defs) and the legacy-symbol re-export surface, plus exercises
each new sibling through the public ``SaveData.save()`` /
``LoadData.load()`` API so the package-private contract stays honest.

LOC outcome:
- ``savers/save_data.py``: 1247 → 618 LOC.
- ``loaders/load_data.py``: 1217 → 733 LOC.
- Both files removed from ``GOD_FILE_SIZE_WAIVERS`` in
  ``scripts/check_god_files.py``. Only the three Bucket D waivers
  (scheduler, registry, git_engine) remain. ``python scripts/check_god_files.py
  --enforce`` reports 0 NEW violations.

Doc updates per P3#1 lesson from Phase 1 audit:
- ``docs/architecture/PROJECT_TREE.md`` — loaders/ + savers/ trees
  now enumerate the new sibling files.
- ``docs/architecture/ARCHITECTURE_legacy.md`` §1336-1337 prose adds a
  pointer to the Path-D sibling layout for each block.
- ``docs/specs/phase11-implementation-standards.md`` T-TRK-007 / T-TRK-008
  "Files to be created" sections note that the code sketches are
  illustrative; the canonical layout uses the private siblings.
- ``docs/adr/ADR-028.md`` ``governs.files`` swapped the single-file
  loader/saver paths for ``src/scistudio/blocks/io/{loaders,savers}/**``
  globs so the new sibling files stay under ADR-028 governance.

Tests run:
- ``pytest tests/blocks/io/`` — 320 pass (was 274 before; 46 new layout
  tests added).
- ``pytest tests/contracts/test_io_capability_contract.py
  tests/engine/test_materialisation.py
  tests/blocks/app/test_bridge_prepare_materialisation.py`` — 26 pass
  + 2 xfail (unchanged pre-existing #1454 drift).
- ``pytest packages/scistudio-blocks-imaging/tests/test_save_image_browse_mode.py``
  — 6 pass (verifies the legacy ``from scistudio.blocks.io.savers.save_data
  import SaveData`` plugin import path still works).
- ``ruff check`` + ``ruff format --check`` — clean.
- ``python scripts/check_god_files.py --enforce`` — 3 waived (Bucket D
  only), 0 new.
- ``python -m scistudio.qa.audit.full_audit`` — status=pass.
- Sentrux scan / check_rules / health — quality_signal 4430 vs Phase 1
  baseline 4441; rules pass.

Owner authorized ``admin-approved:core-change`` label for Phase 2 on
2026-05-22 chat. Path D was approved the same day. No ADR change
required because §C9 is satisfied verbatim by the per-class layout.

Gate-Record: .workflow/records/1459-io-data-helpers.json
Task-Kind: refactor
Issue: #1459
