+++ .workflow/records/1459-io-data-helpers.json
@@ -1,0 +1,171 @@
+{
+  "admin_labels": [],
+  "amendments": [],
+  "branch": "refactor/issue-1459/io-data-helpers",
+  "changed_test_paths": [
+    "tests/blocks/io/test_io_data_helpers_layout.py"
+  ],
+  "check_results": [
+    {
+      "command_or_tool": "ruff check",
+      "exit_code": 0,
+      "name": "ruff",
+      "output_path": null,
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "ruff format --check",
+      "exit_code": 0,
+      "name": "format",
+      "output_path": null,
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "pytest tests/blocks/io/ --timeout=60",
+      "exit_code": 0,
+      "name": "pytest",
+      "output_path": null,
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "scripts/check_god_files.py --enforce",
+      "exit_code": 0,
+      "name": "god_file_advisory",
+      "output_path": null,
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    }
+  ],
+  "commit": null,
+  "docs_landing": {
+    "changelog": {},
+    "checklist": {},
+    "docs": {
+      "paths": [
+        "docs/architecture/PROJECT_TREE.md",
+        "docs/architecture/ARCHITECTURE_legacy.md",
+        "docs/specs/phase11-implementation-standards.md",
+        "docs/adr/ADR-028.md"
+      ]
+    }
+  },
+  "full_audit": {
+    "blocks_merge": false,
+    "command": "python -m scistudio.qa.audit.full_audit",
+    "exit_code": 0,
+    "known_debt": [],
+    "output_path": "docs/audit/full-audit-latest.json",
+    "status": "pass",
+    "summary": {},
+    "unclassified_failures": []
+  },
+  "governance_touch": false,
+  "issues": [
+    {
+      "close_in_pr": true,
+      "followup_rationale": null,
+      "number": 1459,
+      "url": null
+    }
+  ],
+  "owner_directive": "Phase 2 of #1427: extract save_data/load_data helpers to private sibling modules per Path D; no ADR change, \u00a7C9 satisfied verbatim",
+  "planned_files": [
+    "src/scistudio/blocks/io/savers/save_data.py",
+    "src/scistudio/blocks/io/savers/_capability.py",
+    "src/scistudio/blocks/io/savers/_helpers.py",
+    "src/scistudio/blocks/io/savers/_streaming.py",
+    "src/scistudio/blocks/io/loaders/load_data.py",
+    "src/scistudio/blocks/io/loaders/_capability.py",
+    "src/scistudio/blocks/io/loaders/_helpers.py",
+    "tests/blocks/io/test_io_data_helpers_layout.py",
+    "scripts/check_god_files.py",
+    "docs/architecture/PROJECT_TREE.md",
+    "docs/architecture/ARCHITECTURE_legacy.md",
+    "docs/specs/phase11-implementation-standards.md",
+    "docs/adr/ADR-028.md"
+  ],
+  "pull_request": null,
+  "record_path": ".workflow/records/1459-io-data-helpers.json",
+  "required_checks": [
+    "ruff",
+    "format",
+    "pytest",
+    "god_file_advisory",
+    "full_audit",
+    "sentrux"
+  ],
+  "schema_version": "1",
+  "scope": {
+    "exclude": [],
+    "include": [
+      "src/scistudio/blocks/io/savers/**",
+      "src/scistudio/blocks/io/loaders/**",
+      "tests/blocks/io/**",
+      "scripts/check_god_files.py",
+      "docs/architecture/PROJECT_TREE.md",
+      "docs/architecture/ARCHITECTURE_legacy.md",
+      "docs/specs/phase11-implementation-standards.md",
+      "docs/adr/ADR-028.md"
+    ]
+  },
+  "sentrux": {
+    "command_or_tool": "mcp__plugin_sentrux_sentrux__{scan,check_rules,health}",
+    "mode": "free-tier",
+    "name": "sentrux.free_tier",
+    "output_path": null,
+    "pro_required": false,
+    "quality_signal": 4430.0,
+    "rules_checked": 3,
+    "status": "pass",
+    "summary": {},
+    "thresholds": {},
+    "total_rules_defined": 15
+  },
+  "stages": [
+    {
+      "completed_at": null,
+      "stage": "scope_and_issue",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "plan",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "implement",
+      "status": "pending",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "update_docs",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "test_and_checks",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "commit_and_submit_pr",
+      "status": "pending",
+      "summary": null
+    }
+  ],
+  "task_id": "1459-io-data-helpers",
+  "task_kind": "refactor"
+}
