+++ docs/adr/ADR-028.md
@@ -41,8 +41,17 @@ governs:
   files:
     - docs/adr/ADR-028.md
     - src/scistudio/blocks/io/io_block.py
-    - src/scistudio/blocks/io/loaders/load_data.py
-    - src/scistudio/blocks/io/savers/save_data.py
+    # Path-D layout (issue #1459, Phase 2 of umbrella #1427): the
+    # ``LoadData`` / ``SaveData`` classes still own their respective
+    # ``load_data.py`` / ``save_data.py`` module, with module-level
+    # helper functions extracted to private sibling modules under each
+    # sub-package. The glob below covers the canonical class files plus
+    # the new ``_capability.py`` / ``_helpers.py`` / ``_streaming.py``
+    # siblings. §C9 ("private functions, not helper classes") is
+    # satisfied verbatim because every sibling holds module-level
+    # ``_underscore`` functions only.
+    - src/scistudio/blocks/io/loaders/**
+    - src/scistudio/blocks/io/savers/**
     - src/scistudio/blocks/base/block.py
     - src/scistudio/blocks/registry.py
     - src/scistudio/api/schemas.py
@@ -51,6 +60,7 @@ governs:
     - tests/workflow/test_validator_dynamic_ports.py
     - tests/blocks/io/test_load_data.py
     - tests/blocks/io/test_save_data.py
+    - tests/blocks/io/test_io_data_helpers_layout.py
   excludes: []
 tests:
   - tests/blocks/test_dynamic_ports.py
