+++ docs/architecture/ARCHITECTURE_legacy.md
@@ -1333,8 +1333,8 @@ Handles all data ingress and egress. **`IOBlock` is an abstract base class** (AD
 
 Core ships **two concrete IO blocks** that together cover all six core `DataObject` types (`Array`, `DataFrame`, `Series`, `Text`, `Artifact`, `CompositeData`):
 
-- **`LoadData`** at `src/scistudio/blocks/io/loaders/load_data.py` — input-only. Implements `load()` by dispatching to one of six private module-level `_load_*` functions selected by the `core_type` config enum (ADR-028 Addendum 1 §C9).
-- **`SaveData`** at `src/scistudio/blocks/io/savers/save_data.py` — output-only. Mirrors `LoadData` with six private `_save_*` dispatch functions.
+- **`LoadData`** at `src/scistudio/blocks/io/loaders/load_data.py` — input-only. Implements `load()` by dispatching to one of six private module-level `_load_*` functions selected by the `core_type` config enum (ADR-028 Addendum 1 §C9). Module-level helpers were extracted to private sibling modules in #1459 (Phase 2 of #1427): `loaders/_capability.py` (FormatCapability declarations + format dispatch) and `loaders/_helpers.py` (path + pickle helpers, MIME/text format lookups). The siblings hold module-level functions only — §C9 ("private functions, not helper classes") is satisfied verbatim.
+- **`SaveData`** at `src/scistudio/blocks/io/savers/save_data.py` — output-only. Mirrors `LoadData` with six private `_save_*` dispatch functions. Same Path-D sibling layout: `savers/_capability.py`, `savers/_helpers.py`, plus `savers/_streaming.py` for the ADR-031 Phase 3 streaming export paths (zarr-to-zarr copy + CSV/Parquet row-group writers).
 
 Both classes use the **dynamic-port mechanism** (ADR-028 Addendum 1 §C5): they declare a `dynamic_ports: ClassVar[dict[str, Any] | None]` descriptor mapping the `core_type` config field to a per-`enum`-value list of `accepted_types`, and override `get_effective_input_ports()` / `get_effective_output_ports()` so the validator and the frontend palette see the *narrowed* type for the configured `core_type`. The static `output_ports` declaration uses the broad `[DataObject]` upper bound for backward-compatible registration; the per-instance override is what the runtime and the GUI actually consume.
