+++ src/scistudio/blocks/io/savers/save_data.py
@@ -29,12 +29,35 @@ ADR-043 / spec ``adr-043-package-migration`` FR-002 / FR-003:
 been removed; format dispatch is derived from the capability
 declarations via :func:`_legacy_save_extension_map`. Each save
 capability is paired with its load sibling via ``roundtrip_group``.
+
+Module layout (Path D, issue #1459, Phase 2 of #1427)
+-----------------------------------------------------
+
+The module-level helper functions that used to live in this file were
+extracted to private sibling modules per ADR-028 Addendum 1 §C9
+("private functions, not helper classes" — Path D satisfies this
+verbatim because the new siblings contain *only* module-level
+``_underscore`` functions, no classes):
+
+- :mod:`scistudio.blocks.io.savers._capability` — ``FormatCapability``
+  declarations, ``_SAVE_EXTENSION_MAP``, ``_resolve_save_format``.
+- :mod:`scistudio.blocks.io.savers._helpers` — ``_require_path``,
+  ``_check_pickle_gate``, ``_unwrap_for_save``, ``_CORE_TYPE_MAP``,
+  and the small per-format coercion helpers shared by the dispatch
+  functions.
+- :mod:`scistudio.blocks.io.savers._streaming` — row-group / zero
+  materialisation export paths used by ``_save_array`` and
+  ``_save_dataframe`` (ADR-031 Phase 3).
+
+The helper symbols are re-imported at module level below so legacy
+callers that do ``from scistudio.blocks.io.savers.save_data import
+_SAVE_CAPABILITIES`` (test code, mostly) continue to work without
+change.
 """
 
 from __future__ import annotations
 
 import json
-import logging
 import pickle
 import shutil
 from pathlib import Path
@@ -42,8 +65,38 @@ from typing import Any, ClassVar
 
 from scistudio.blocks.base.config import BlockConfig
 from scistudio.blocks.base.ports import InputPort, OutputPort
-from scistudio.blocks.io.capabilities import FormatCapability, MetadataFidelity
+from scistudio.blocks.io.capabilities import FormatCapability
 from scistudio.blocks.io.io_block import IOBlock
+
+# Re-exports from the private sibling modules. The unused-import noqa
+# tags are intentional: legacy callers import these names directly from
+# :mod:`scistudio.blocks.io.savers.save_data`, so they must remain
+# present at this module's top level.
+from scistudio.blocks.io.savers._capability import (
+    _PICKLE_NOTE,  # noqa: F401  re-export for backward compat
+    _SAVE_CAPABILITIES,
+    _SAVE_EXTENSION_MAP,
+    _legacy_save_extension_map,  # noqa: F401  re-export for backward compat
+    _resolve_save_format,
+    _save_capability,  # noqa: F401  re-export for backward compat
+    _supported_save_extensions,
+)
+from scistudio.blocks.io.savers._helpers import (
+    _CORE_TYPE_MAP,
+    _PICKLE_EXTENSIONS,  # noqa: F401  re-export for backward compat
+    _check_pickle_gate,
+    _dataframe_to_arrow_table,
+    _require_path,
+    _resolve_core_type_name,
+    _slot_path_for,
+    _unwrap_for_save,
+    logger,  # noqa: F401  re-export for backward compat
+)
+from scistudio.blocks.io.savers._streaming import (
+    _streaming_save_dataframe_csv,
+    _streaming_save_dataframe_parquet,
+    _zarr_store_copy,
+)
 from scistudio.core.types.array import Array
 from scistudio.core.types.artifact import Artifact
 from scistudio.core.types.base import DataObject
@@ -52,579 +105,6 @@ from scistudio.core.types.composite import CompositeData
 from scistudio.core.types.dataframe import DataFrame
 from scistudio.core.types.series import Series
 from scistudio.core.types.text import Text
-
-logger = logging.getLogger(__name__)
-
-
-# ADR-028 Addendum 1 §C5: hardcoded six core types. The keys are the
-# string enum values exposed in ``config_schema['properties']['core_type']
-# ['enum']``; the values are the concrete :class:`DataObject` subclasses
-# used by :meth:`SaveData.get_effective_input_ports` to update the
-# accepted-types of the ``data`` input port.
-_CORE_TYPE_MAP: dict[str, type[DataObject]] = {
-    "Array": Array,
-    "DataFrame": DataFrame,
-    "Series": Series,
-    "Text": Text,
-    "Artifact": Artifact,
-    "CompositeData": CompositeData,
-}
-
-
-# Pickle file extensions that require explicit ``allow_pickle=True``
-# opt-in. Treated case-insensitively.
-_PICKLE_EXTENSIONS: frozenset[str] = frozenset({".pkl", ".pickle"})
-
-
-# ---------------------------------------------------------------------------
-# ADR-043 / spec ``adr-043-package-migration`` FR-002:
-# explicit ``FormatCapability`` declarations for the SaveData six-core-type
-# matrix. Capability id convention: ``core.{lower(type)}.{format_id}.save``.
-# Each capability is paired with its load sibling (declared in
-# ``scistudio.blocks.io.loaders.load_data``) via
-# ``roundtrip_group=core.{type}.{format}``.
-# ---------------------------------------------------------------------------
-
-
-_PICKLE_NOTE: str = "requires allow_pickle=True"
-
-
-def _save_capability(
-    *,
-    data_type: type[DataObject],
-    type_name: str,
-    format_id: str,
-    extensions: tuple[str, ...],
-    label: str,
-    handler: str,
-    notes: str | None = None,
-) -> FormatCapability:
-    """Build a single save-direction :class:`FormatCapability` record.
-
-    The capability id follows the spec FR-015 convention
-    ``core.{lower(type)}.{format_id}.save`` and the roundtrip group
-    mirrors the matching load capability so the registry can pair
-    load+save handlers via :attr:`FormatCapability.roundtrip_group`.
-
-    Core IO records are declared ``is_default=False`` so installed
-    package-specific savers (e.g. the LCMS package's
-    ``scistudio-blocks-lcms.table.csv.save`` for ``(DataFrame, .csv)``)
-    keep ownership of the default slot for their specialised data types
-    without triggering a registration-time conflict. When no package
-    declares a default for a given ``(data_type, extension)`` slot, the
-    registry returns the unique non-default core capability normally per
-    :meth:`BlockRegistry.find_saver_capability`.
-    """
-
-    lower_type = type_name.lower()
-    return FormatCapability(
-        id=f"core.{lower_type}.{format_id}.save",
-        direction="save",
-        data_type=data_type,
-        format_id=format_id,
-        extensions=extensions,
-        label=label,
-        block_type="SaveData",
-        handler=handler,
-        is_default=False,
-        roundtrip_group=f"core.{lower_type}.{format_id}",
-        metadata_fidelity=MetadataFidelity(level="pixel_only", notes=notes),
-        is_synthesized=False,
-    )
-
-
-_SAVE_CAPABILITIES: tuple[FormatCapability, ...] = (
-    # ----- Array -----------------------------------------------------------
-    _save_capability(
-        data_type=Array,
-        type_name="Array",
-        format_id="npy",
-        extensions=(".npy",),
-        label="NumPy NPY",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Array,
-        type_name="Array",
-        format_id="npz",
-        extensions=(".npz",),
-        label="NumPy NPZ",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Array,
-        type_name="Array",
-        format_id="zarr",
-        extensions=(".zarr",),
-        label="Zarr",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Array,
-        type_name="Array",
-        format_id="parquet",
-        extensions=(".parquet", ".pq"),
-        label="Parquet",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Array,
-        type_name="Array",
-        format_id="pickle",
-        extensions=(".pkl", ".pickle"),
-        label="Pickle",
-        handler="save",
-        notes=_PICKLE_NOTE,
-    ),
-    # ----- DataFrame -------------------------------------------------------
-    _save_capability(
-        data_type=DataFrame,
-        type_name="DataFrame",
-        format_id="csv",
-        extensions=(".csv",),
-        label="CSV",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=DataFrame,
-        type_name="DataFrame",
-        format_id="tsv",
-        extensions=(".tsv",),
-        label="TSV",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=DataFrame,
-        type_name="DataFrame",
-        format_id="parquet",
-        extensions=(".parquet", ".pq"),
-        label="Parquet",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=DataFrame,
-        type_name="DataFrame",
-        format_id="json",
-        extensions=(".json",),
-        label="JSON",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=DataFrame,
-        type_name="DataFrame",
-        format_id="pickle",
-        extensions=(".pkl", ".pickle"),
-        label="Pickle",
-        handler="save",
-        notes=_PICKLE_NOTE,
-    ),
-    # ----- Series ----------------------------------------------------------
-    _save_capability(
-        data_type=Series,
-        type_name="Series",
-        format_id="csv",
-        extensions=(".csv",),
-        label="CSV",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Series,
-        type_name="Series",
-        format_id="tsv",
-        extensions=(".tsv",),
-        label="TSV",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Series,
-        type_name="Series",
-        format_id="parquet",
-        extensions=(".parquet", ".pq"),
-        label="Parquet",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Series,
-        type_name="Series",
-        format_id="json",
-        extensions=(".json",),
-        label="JSON",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Series,
-        type_name="Series",
-        format_id="pickle",
-        extensions=(".pkl", ".pickle"),
-        label="Pickle",
-        handler="save",
-        notes=_PICKLE_NOTE,
-    ),
-    # ----- Text ------------------------------------------------------------
-    # ``.markdown`` and ``.htm`` are accepted by ``_save_text`` (see the
-    # Text-specific superset in that function) but were missing from the
-    # registry-visible capability declaration, so ``find_saver_capability``
-    # could not discover SaveData for those extensions (#1110).
-    _save_capability(
-        data_type=Text,
-        type_name="Text",
-        format_id="text",
-        extensions=(
-            ".txt",
-            ".log",
-            ".md",
-            ".markdown",
-            ".html",
-            ".htm",
-            ".xml",
-            ".yaml",
-            ".yml",
-            ".toml",
-        ),
-        label="Text",
-        handler="save",
-    ),
-    # Text + .json is a legacy save-only extension: ``_save_text`` writes
-    # JSON files happily (just dumps ``obj.content``) but LoadData's Text
-    # capability intentionally excludes ``.json`` because ``_load_text``
-    # does not parse JSON. Declared as a separate capability record with
-    # ``format_id="json"`` so ``find_saver_capability(Text, '.json')``
-    # resolves to a Text saver without conflicting with the Text +
-    # ``.json`` -> ``"text"`` mapping at the extension-map layer. Mirrors
-    # the Series + json save-only legacy branch and the P1 Codex-review
-    # finding on PR #1300.
-    _save_capability(
-        data_type=Text,
-        type_name="Text",
-        format_id="json",
-        extensions=(".json",),
-        label="Text as JSON",
-        handler="save",
-    ),
-    # ----- Artifact (opaque catch-all saver) ------------------------------
-    # ``_save_artifact`` accepts any extension at runtime — it copies bytes
-    # to the configured path. The capability records below cover the
-    # canonical MIME-mapped extensions AND every extension the typed core
-    # savers claim, so the AppBlock wildcard-port flow (which the binner
-    # maps to Artifact) can resolve a saver for the legacy
-    # supported-extension union without falling through to the
-    # registry's polymorphic fallback. Mirror of the corresponding
-    # ``Artifact``-as-opaque-loader records on the ``LoadData`` side.
-    _save_capability(
-        data_type=Artifact,
-        type_name="Artifact",
-        format_id="binary",
-        extensions=(".bin", ".dat"),
-        label="Binary",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Artifact,
-        type_name="Artifact",
-        format_id="pdf",
-        extensions=(".pdf",),
-        label="PDF",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Artifact,
-        type_name="Artifact",
-        format_id="png",
-        extensions=(".png",),
-        label="PNG",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Artifact,
-        type_name="Artifact",
-        format_id="jpeg",
-        extensions=(".jpg", ".jpeg"),
-        label="JPEG",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Artifact,
-        type_name="Artifact",
-        format_id="tiff",
-        extensions=(".tif", ".tiff"),
-        label="TIFF",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Artifact,
-        type_name="Artifact",
-        format_id="csv",
-        extensions=(".csv",),
-        label="CSV (opaque)",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Artifact,
-        type_name="Artifact",
-        format_id="tsv",
-        extensions=(".tsv",),
-        label="TSV (opaque)",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Artifact,
-        type_name="Artifact",
-        format_id="json",
-        extensions=(".json",),
-        label="JSON (opaque)",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Artifact,
-        type_name="Artifact",
-        format_id="parquet",
-        extensions=(".parquet", ".pq"),
-        label="Parquet (opaque)",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Artifact,
-        type_name="Artifact",
-        format_id="npy",
-        extensions=(".npy",),
-        label="NumPy NPY (opaque)",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Artifact,
-        type_name="Artifact",
-        format_id="npz",
-        extensions=(".npz",),
-        label="NumPy NPZ (opaque)",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Artifact,
-        type_name="Artifact",
-        format_id="zarr",
-        extensions=(".zarr",),
-        label="Zarr (opaque)",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Artifact,
-        type_name="Artifact",
-        format_id="text",
-        extensions=(
-            ".txt",
-            ".log",
-            ".md",
-            ".markdown",
-            ".html",
-            ".htm",
-            ".xml",
-            ".yaml",
-            ".yml",
-            ".toml",
-        ),
-        label="Text (opaque)",
-        handler="save",
-    ),
-    _save_capability(
-        data_type=Artifact,
-        type_name="Artifact",
-        format_id="pickle",
-        extensions=(".pkl", ".pickle"),
-        label="Pickle (opaque)",
-        handler="save",
-        notes=_PICKLE_NOTE,
-    ),
-    # ----- CompositeData ---------------------------------------------------
-    _save_capability(
-        data_type=CompositeData,
-        type_name="CompositeData",
-        format_id="json",
-        extensions=(".json",),
-        label="Composite manifest (JSON)",
-        handler="save",
-    ),
-)
-
-
-def _legacy_save_extension_map(
-    capabilities: tuple[FormatCapability, ...],
-) -> dict[str, str]:
-    """Derive a legacy ``extension -> format_id`` mapping from capabilities.
-
-    Mirror of :func:`scistudio.blocks.io.loaders.load_data._legacy_extension_map`.
-    The same extension may appear on multiple types (e.g. ``.parquet`` on
-    Array, DataFrame, Series) but the ``format_id`` is identical across
-    types, so the dict is well-defined. Conflicting format_id values
-    raise :class:`RuntimeError` to surface misconfiguration.
-    """
-
-    mapping: dict[str, str] = {}
-    for capability in capabilities:
-        for extension in capability.extensions:
-            normalized = extension.lower()
-            existing = mapping.get(normalized)
-            if existing is not None and existing != capability.format_id:
-                raise RuntimeError(
-                    f"SaveData format_capabilities declare conflicting format_id "
-                    f"for extension {normalized!r}: {existing!r} vs {capability.format_id!r}"
-                )
-            mapping[normalized] = capability.format_id
-    return mapping
-
-
-_SAVE_EXTENSION_MAP: dict[str, str] = _legacy_save_extension_map(_SAVE_CAPABILITIES)
-
-
-def _supported_save_extensions() -> tuple[str, ...]:
-    """Return the sorted tuple of every extension declared by SaveData.
-
-    Replaces the legacy ``sorted(SaveData.supported_extensions.keys())``
-    call sites in user-facing error messages.
-    """
-
-    return tuple(sorted(_SAVE_EXTENSION_MAP.keys()))
-
-
-def _resolve_save_format(path: Path) -> str | None:
-    """Resolve a path's format identifier from the SaveData capability map.
-
-    ADR-043 / spec FR-003: format dispatch is now derived from explicit
-    :attr:`SaveData.format_capabilities` rather than the deleted
-    ``supported_extensions`` ClassVar. Compound-suffix-first,
-    case-insensitive lookup matching :meth:`IOBlock._detect_format`
-    semantics.
-    """
-
-    if not _SAVE_EXTENSION_MAP:
-        return None
-    suffixes = [s.lower() for s in path.suffixes]
-    for start in range(len(suffixes)):
-        candidate = "".join(suffixes[start:])
-        if candidate in _SAVE_EXTENSION_MAP:
-            return _SAVE_EXTENSION_MAP[candidate]
-    return None
-
-
-def _require_path(config: BlockConfig) -> Path:
-    """Return the configured ``path`` as a :class:`Path` or raise.
-
-    Centralised so each ``_save_*`` function can fail loudly with the
-    same message instead of duplicating the validation.
-    """
-    raw = config.get("path")
-    if raw is None or raw == "":
-        raise ValueError("SaveData requires a non-empty 'path' in config.params.")
-    path = Path(str(raw))
-    path.parent.mkdir(parents=True, exist_ok=True)
-    return path
-
-
-def _check_pickle_gate(path: Path, config: BlockConfig) -> bool:
-    """Return ``True`` if *path* is a pickle file.
-
-    Raises :class:`ValueError` if the file extension is ``.pkl`` /
-    ``.pickle`` and ``allow_pickle`` is not explicitly ``True``. When
-    pickle is enabled, an explicit security warning is logged at
-    ``WARNING`` level so the user can audit the workflow run.
-    """
-    if path.suffix.lower() not in _PICKLE_EXTENSIONS:
-        return False
-    if not bool(config.get("allow_pickle", False)):
-        raise ValueError(
-            f"Refusing to write pickle file {path.name!r}: pickle is opt-in for "
-            "security reasons. Set 'allow_pickle': True in the block config to "
-            "enable pickle writes."
-        )
-    logger.warning(
-        "SaveData: writing pickle file %r — pickle files can execute arbitrary "
-        "code on load and should only be used with trusted data.",
-        str(path),
-    )
-    return True
-
-
-# ---------------------------------------------------------------------------
-# ADR-031 Phase 3 (Task 18): Streaming export helpers
-# ---------------------------------------------------------------------------
-
-
-def _zarr_store_copy(src_path: str, dst_path: str) -> None:
-    """Copy a zarr store from *src_path* to *dst_path* without materialisation.
-
-    Uses ``shutil.copytree`` for a file-level copy of the zarr directory
-    store. This copies the compressed chunks directly, avoiding any
-    decompression/recompression round-trip.
-    """
-    dst = Path(dst_path)
-    if dst.exists():
-        shutil.rmtree(dst)
-    shutil.copytree(src_path, dst_path)
-
-
-def _streaming_save_dataframe_csv(obj: DataObject, path: Path, delimiter: str = ",") -> None:
-    """Stream a storage-backed DataFrame to CSV/TSV via row-group batches.
-
-    Reads from the arrow backend in chunks and writes each batch to the
-    output file, avoiding full materialisation of the entire table.
-    """
-    import pyarrow.csv as pcsv
-
-    ref = getattr(obj, "_storage_ref", None)
-    if ref is None or ref.backend != "arrow":
-        # Fallback: full materialisation
-        table = _dataframe_to_arrow_table(obj)  # type: ignore[arg-type]
-        pcsv.write_csv(
-            table,
-            str(path),
-            write_options=pcsv.WriteOptions(delimiter=delimiter),
-        )
-        return
-
-    from scistudio.core.storage.arrow_backend import ArrowBackend
-
-    backend = ArrowBackend()
-    # Write header from first chunk, then append remaining chunks
-    first_chunk = True
-    with open(str(path), "wb") as fh:
-        for chunk_table in backend.iter_chunks(ref, chunk_size=65536):
-            pcsv.write_csv(
-                chunk_table,
-                fh,
-                write_options=pcsv.WriteOptions(
-                    delimiter=delimiter,
-                    include_header=first_chunk,
-                ),
-            )
-            first_chunk = False
-
-
-def _streaming_save_dataframe_parquet(obj: DataObject, path: Path) -> None:
-    """Stream a storage-backed DataFrame to Parquet via row-group batches.
-
-    Reads from the arrow backend in chunks and writes each batch as a
-    separate row group, avoiding full materialisation.
-    """
-    import pyarrow.parquet as pq
-
-    ref = getattr(obj, "_storage_ref", None)
-    if ref is None or ref.backend != "arrow":
-        # Fallback: full materialisation
-        table = _dataframe_to_arrow_table(obj)  # type: ignore[arg-type]
-        pq.write_table(table, str(path))
-        return
-
-    from scistudio.core.storage.arrow_backend import ArrowBackend
-
-    backend = ArrowBackend()
-    writer = None
-    try:
-        for chunk_table in backend.iter_chunks(ref, chunk_size=65536):
-            if writer is None:
-                writer = pq.ParquetWriter(str(path), chunk_table.schema)
-            writer.write_table(chunk_table)
-    finally:
-        if writer is not None:
-            writer.close()
 
 
 class SaveData(IOBlock):
@@ -778,50 +258,6 @@ class SaveData(IOBlock):
             "CompositeData": _save_composite_data,
         }
         dispatch[type_name](unwrapped, config)
-
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-
-def _unwrap_for_save(
-    obj: DataObject | Collection,
-    target_cls: type[DataObject],
-) -> DataObject:
-    """Unwrap a single-item :class:`Collection` or return the bare object.
-
-    Per spec §j, mixed-type Collections (where some item is not a
-    ``target_cls`` instance) raise :class:`ValueError`. A Collection
-    of all-``target_cls`` items with exactly one element is unwrapped
-    transparently. A Collection of length > 1 also raises because the
-    save dispatch functions write a single file at the configured path.
-    """
-    if isinstance(obj, Collection):
-        items = list(obj)
-        if not items:
-            raise ValueError("SaveData received an empty Collection; nothing to save.")
-        if len(items) > 1:
-            raise ValueError(
-                f"SaveData received a Collection of {len(items)} items; "
-                "core SaveData writes one file at the configured path. "
-                "Iterate over the Collection upstream and call SaveData per item."
-            )
-        only = items[0]
-        if not isinstance(only, target_cls):
-            raise ValueError(
-                f"SaveData(core_type={target_cls.__name__}) received a "
-                f"Collection item of type {type(only).__name__}; mixed-type "
-                "Collections are out-of-scope per ADR-028 Addendum 1 §C9."
-            )
-        return only
-    if not isinstance(obj, target_cls):
-        raise ValueError(
-            f"SaveData(core_type={target_cls.__name__}) received an instance of "
-            f"{type(obj).__name__}; the input must be a {target_cls.__name__} "
-            "instance (or a single-item Collection thereof)."
-        )
-    return obj
 
 
 # ---------------------------------------------------------------------------
@@ -1167,71 +603,6 @@ def _save_composite_data(obj: DataObject, config: BlockConfig) -> None:
         json.dumps(manifest, ensure_ascii=False, indent=2),
         encoding="utf-8",
     )
-
-
-# ---------------------------------------------------------------------------
-# Internal helpers shared by the dispatch functions
-# ---------------------------------------------------------------------------
-
-
-def _dataframe_to_arrow_table(obj: DataFrame) -> Any:
-    """Coerce a :class:`DataFrame` into a :class:`pyarrow.Table`.
-
-    ADR-031 D6: always routes through ``get_in_memory_data()`` ->
-    ``to_memory()`` -> storage backend read. The former ``_arrow_table``
-    backdoor is removed.
-    """
-    import pyarrow as pa
-
-    raw = obj.get_in_memory_data()
-    if isinstance(raw, pa.Table):
-        return raw
-    if isinstance(raw, dict):
-        return pa.table(raw)
-    if isinstance(raw, list):
-        return pa.Table.from_pylist(raw)
-    raise ValueError(
-        f"Cannot convert DataFrame in-memory data of type {type(raw).__name__} "
-        "to a pyarrow.Table; expected pa.Table, dict, or list of records."
-    )
-
-
-def _resolve_core_type_name(obj: DataObject) -> str | None:
-    """Map a :class:`DataObject` instance to its core-type enum name.
-
-    Returns ``None`` if *obj* is not an instance of any of the six
-    supported core types. Order matters here: more specific types
-    (``Array``, ``CompositeData``) come before the base ``DataObject``
-    fallback.
-    """
-    for type_name, cls in _CORE_TYPE_MAP.items():
-        if isinstance(obj, cls):
-            return type_name
-    return None
-
-
-def _slot_path_for(
-    type_name: str,
-    slots_dir: Path,
-    slot_name: str,
-) -> tuple[str, Path]:
-    """Return ``(filename, full_path)`` for a CompositeData slot file.
-
-    The default extension per type mirrors the most common round-trip
-    format used by tests: ``.csv`` for DataFrame / Series, ``.npy``
-    for Array, ``.txt`` for Text, ``.bin`` for Artifact, ``.json`` for
-    nested CompositeData.
-    """
-    default_ext = {
-        "Array": ".npy",
-        "DataFrame": ".csv",
-        "Series": ".csv",
-        "Text": ".txt",
-        "Artifact": ".bin",
-        "CompositeData": ".json",
-    }[type_name]
-    filename = f"{slot_name}{default_ext}"
-    return filename, slots_dir / filename
 
 
 # Slot dispatch table — separate from the main dispatch so that
